Add audio processing module
The base audio processing provided by the web audio api is not enough
for this project, so I must implement my own audio processing nodes.
This should be fast and reliable and run off the main thread, so I will
be implementing it in rust and building it to wasm.