Add audio engine to build pipeline
The audio processing that will not be handled by the base objects in the
web audio api will be done in a wasm module.

Create-react-app does not support either: the packing of rust to wasm
modules, nor the loading of wasm into the final webpack module. Because
of this, I have switched to react-app-rewired so that I can add the
loader for wasm code, and use the wasm pack plugin for webpack that will
build the wasm module from the rust code.

The build module is then included as a package in the dependency list so
that any imports are easily readable.