Implement VCO1 processor
The processor is in the public folder instead of the src directory
because the module is loaded in a worker context instead of being
imported. This means that we can do one of two things:

- Implement a babel/webpack loader to identify these modules and then
  compile them separately to the main code. (Similarly to how
  babel/webpack handles normal workers)

- Leave the code un-minified and load it in production as it was
  written.

The second option is what I have decided to go with for now. It means I
don't have to worry about making changes to the configs under the hood
of create react app. The drawback is that the final bundle will not be
as small as it could be. This will hopefully be remedied in the future
with the use of web assembly; the [currently javascript] modules in the
public folder will be implemented in rust and compiled to web assembly,
making the final bundle only as large as it has to be while potentially
increasing performance.