Implement friendly range slider
The base HTML range slider is not useful for turning knobs as it will
jump to wherever is clicked on the track. The new range slider will not
jump, but change based on the difference in displacement from the
initial click, making it much easier to make fine adjustments and avoid
large jumps.

The touch input is not yet implemented.