Factor value range into drag resistance
The resistance to dragging was being calculated as a multiple of the
angles that the knob would go through. This failed to make each knob
feel natural as knobs with different value ranges would have different
resistances.

The resulting resistance was a function of the range only and thus could
be implemented solely by the range slider.