Abstact state selectors
The selectors for the state object in the observe store methods were set
in concrete. Now they are able to be set in the constructor.

This cuts down on duplicate code between similar components. For
example, both the filters have frequency modulation on their cutoff
frequencies. By setting the selectors differently, we can reuse the
frequency modulator node instead of rewriting the entire node again just
to change how it connects tothe state.