Factor out code into wrapper
There were many knobs with a value from 0 to 1 and which had a entire
module with map state and dispatch to props where the only change was in
the part of the state the knob was watching and the action creator it
used to update the value in the state.

The new ValueKnob container fixes this.