Connect Keyboard to state object
Up until now I was under the impression that the keyboard should not be
part of the state object, as only note on/off changes were important and
they were not persistant. While implementing the keyboard however I
found that I was keeping the state anyway for the ui, and this was just
complicating things. When i went to add the computer keyboard as an
input device, it became too much and I realised that the state of the
keyboard probably is important enough to keep it on the state object.