Prevent mobile access
The web-synth does not work well on mobile devices. This may be due to
the javascript implementations of the oscillators being slow.

By implementing a landing page, desktop users can get a heads up about
how to use the app and mobile users can be alerted about the lack of
support.